Add option to provide predefined server to ginx Server
There are a number of cases in which it is valuable to explicitly
set the Gin engine used by the ginx server. This commit adds a new
type ServerOption that can be used to define options for ginx
servers, as well as a server option to specify an existing Gin
engine.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>